+++ equal.ts
@@ -12,9 +12,11 @@ import {
   isBothPrimitive,
   isBothRegExp,
   isBothSet,
-} from "./is.ts";
-import { entriesSymbol, instanceOf } from "./utils.ts";
-import { is } from "./constants.ts";
+  isBothURL,
+  isBothURLSearchParams,
+} from "./_is.ts";
+import { entriesSymbol, instanceOf } from "./_utils.ts";
+import { is } from "./_constants.ts";
 type Verdict = [
   AnyFn<unknown, readonly [boolean, boolean]>,
   AnyFn<any, boolean>,
@@ -51,6 +53,8 @@ const equal = <T, U extends T>(a: T, b: U): boolean => {
     [isBothError, equalError],
     [isBothMap, equalMap],
     [isBothSet, equalSet],
+    [isBothURL, equalURL],
+    [isBothURLSearchParams, equalURLSearchParams],
     [isBothObjectExcludeJSON, equalObjectExcludeJson],
   ];
 
@@ -123,8 +127,15 @@ const equalKeyValueTupleNoOrder = <T extends [unknown, unknown][], U extends T>(
   a: T,
   b: U,
 ): boolean => {
-  return a.every((tupleA) =>
-    b.some((tupleB) => equalKeyValueTuple(tupleA, tupleB))
+  if (length(a) !== length(b)) return false;
+
+  // TODO: This logic is terrible performance
+  return and(
+    a.every((tupleA) => b.some((tupleB) => equalKeyValueTuple(tupleA, tupleB))),
+    () =>
+      b.every((tupleB) =>
+        a.some((tupleA) => equalKeyValueTuple(tupleA, tupleB))
+      ),
   );
 };
 
@@ -169,6 +180,13 @@ const equalArray = <T extends unknown[], U extends T>(a: T, b: U): boolean => {
   return a.every((val, index) => equal(val, b[index]));
 };
 
+const equalURL = <T extends URL, U extends T>(a: T, b: U): boolean =>
+  a.toString() === b.toString();
+const equalURLSearchParams = <T extends URLSearchParams, U extends T>(
+  a: T,
+  b: U,
+) => equalKeyValueTupleNoOrder([...a], [...b]);
+
 export {
   equal,
   equalArray,
@@ -183,4 +201,6 @@ export {
   equalObjectExcludeJson,
   equalRegExp,
   equalSet,
+  equalURL,
+  equalURLSearchParams,
 };