+++ equal_test.ts
@@ -14,6 +14,8 @@ import {
   equalObjectExcludeJson,
   equalRegExp,
   equalSet,
+  equalURL,
+  equalURLSearchParams,
 } from "./equal.ts";
 
 Deno.test("equalJsonObject", () => {
@@ -89,6 +91,7 @@ Deno.test("equalObjectExcludeJson", () => {
     [new Boolean(false), new Boolean(true), false],
     [new Boolean(true), new Boolean(false), false],
     [new Boolean(false), new Boolean(false), true],
+    [new Object(), new Object(), false],
   ];
 
   table.forEach(([a, b, expected]) => {
@@ -446,6 +449,115 @@ Deno.test("equalDate", () => {
   });
 });
 
+Deno.test("equalURL", () => {
+  const BASE_URL = "https://google.com";
+  const table: [URL, URL, boolean][] = [
+    [new URL(BASE_URL), new URL(BASE_URL), true],
+    [new URL(BASE_URL), new URL(`${BASE_URL}/`), true],
+    [new URL(`${BASE_URL}/`), new URL(BASE_URL), true],
+    [new URL(`${BASE_URL}/`), new URL(`${BASE_URL}/`), true],
+    [new URL(`${BASE_URL}:3000`), new URL(`${BASE_URL}/`), false],
+    [new URL(`${BASE_URL}:3000`), new URL(`${BASE_URL}:80`), false],
+    [new URL(`${BASE_URL}:3000`), new URL(`${BASE_URL}:3000`), true],
+  ];
+  table.forEach(([a, b, expected]) => {
+    assertEquals(
+      equalURL(a, b),
+      expected,
+      `equalURL(${a}, ${b}) -> ${expected}`,
+    );
+  });
+});
+Deno.test("equalURLSearchParams", () => {
+  const table: [URLSearchParams, URLSearchParams, boolean][] = [
+    [new URLSearchParams(), new URLSearchParams(), true],
+    [new URLSearchParams([]), new URLSearchParams(), true],
+    [new URLSearchParams(), new URLSearchParams([]), true],
+    [new URLSearchParams([]), new URLSearchParams([]), true],
+    [new URLSearchParams({}), new URLSearchParams({}), true],
+    [new URLSearchParams({}), new URLSearchParams(), true],
+    [new URLSearchParams(), new URLSearchParams({}), true],
+    [new URLSearchParams({ a: "hello" }), new URLSearchParams(), false],
+    [
+      new URLSearchParams({ a: "hello" }),
+      new URLSearchParams({ a: "hello" }),
+      true,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ a: "hello" }),
+      false,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ a: "hello", b: "world" }),
+      true,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ b: "world", a: "hello" }),
+      true,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ a: "world", b: "hello" }),
+      false,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ a: "world", b: "hello", c: "tom" }),
+      false,
+    ],
+    [
+      new URLSearchParams([["a", "hello"]]),
+      new URLSearchParams([["a", "hello"]]),
+      true,
+    ],
+    [
+      new URLSearchParams([["a", "hello"]]),
+      new URLSearchParams([["a", "world"]]),
+      false,
+    ],
+    [
+      new URLSearchParams([["a", "hello"], ["b", "world"]]),
+      new URLSearchParams([["a", "hello"]]),
+      false,
+    ],
+    [
+      new URLSearchParams([["a", "hello"]]),
+      new URLSearchParams([["a", "hello"], ["b", "world"]]),
+      false,
+    ],
+    [
+      new URLSearchParams([["a", "hello"], ["b", "world"]]),
+      new URLSearchParams([["a", "hello"], ["b", "world"]]),
+      true,
+    ],
+    [
+      new URLSearchParams([["a", "hello"], ["b", "world"]]),
+      new URLSearchParams([["b", "world"], ["a", "hello"]]),
+      true,
+    ],
+    [
+      new URLSearchParams([["a", "hello"]]),
+      new URLSearchParams({ a: "hello" }),
+      true,
+    ],
+    [
+      new URLSearchParams([["b", "world"], ["a", "hello"]]),
+      new URLSearchParams({ a: "hello", b: "world" }),
+      true,
+    ],
+  ];
+  table.forEach(([a, b, expected]) => {
+    assertEquals(
+      equalURLSearchParams(a, b),
+      expected,
+      `equalURLSearchParams(${a}, ${b}) -> ${expected}`,
+    );
+  });
+});
+
 Deno.test("equalFunction", () => {
   const a = () => 1;
   const b = () => 1;
@@ -546,7 +658,7 @@ Deno.test("equalKeyValueTupleNoOrder", () => {
     [[], [], true],
     [[[1, 2], [3, 4]], [[1, 2]], false],
     [[[1, 2], [3, 4]], [[1, 2], [1, 2]], false],
-    [[[1, 2], [1, 2]], [[1, 2], [3, 4]], true],
+    [[[1, 2], [1, 2]], [[1, 2], [3, 4]], false],
     [[[undefined, null], [null, undefined]], [[null, undefined]], false],
     [[[undefined, null], [null, undefined]], [[null, undefined], [
       undefined,
@@ -558,7 +670,7 @@ Deno.test("equalKeyValueTupleNoOrder", () => {
     ]], false],
     [[[1, ""]], [[1, ""]], true],
     [[[1, ""]], [[2, ""]], false],
-    // [[[1, ""]], [[2, ""], [1, ""]], false],
+    [[[1, ""]], [[2, ""], [1, ""]], false],
     [[[1, ""], [2, ""]], [[2, ""], [1, ""]], true],
     [[["", ""]], [["", ""]], true],
     [[["", ""], [1, 2], [null, undefined]], [[null, undefined], ["", ""], [
@@ -829,6 +941,24 @@ Deno.test("equal", () => {
     ],
     [new Set(), new Set(), true],
     [new Set(), new Set([]), true],
+    [new URL("https://google.com"), new URL("https://google.com"), true],
+    [new URL("https://google.com"), new URL("https://google.com:3000"), false],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ a: "hello", b: "world" }),
+      true,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ b: "world", a: "hello" }),
+      true,
+    ],
+    [
+      new URLSearchParams({ a: "hello", b: "world" }),
+      new URLSearchParams({ b: "tom", a: "hello" }),
+      false,
+    ],
+    [new Uint16Array(), new Uint16Array(), false],
   ];
   table.forEach(([a, b, expected]) => {
     assertEquals(